dsadsa

+++ run_tnsm_mif.py
@@ -540,7 +540,7 @@ def main():
                             cmd += " --window_length {}".format(window)
                             cmd += " --input_file_swarm {}".format(dt_failed)
                             cmd += " --save_file_samples {}".format(output_dataset_training_in)
-                            cmd += " --number_blocks {}".format(32*(256/window))
+                            cmd += " --number_blocks {}".format(int(32 * (256 / window)))
                             run_cmd(cmd)
                         output_dataset_training_ins.append((output_dataset_training_in, lr))
                         output_dataset_merged_in += "{};".format(output_dataset_training_in)
@@ -609,7 +609,7 @@ def main():
                     cmd += " --window_length {}".format(window)
                     cmd += " --input_file_swarm {}".format(INPUT_DATASET_TRAINING_OUT)
                     cmd += " --save_file_samples {}".format(output_dataset_training_out)
-                    cmd += " --number_blocks {}".format(32 * (256 / window))
+                    cmd += " --number_blocks {}".format(int(32 * (256 / window)))
 
                     run_cmd(cmd)
 
@@ -654,7 +654,7 @@ def main():
                                 cmd += " --load_samples_out {}".format(output_dataset_training_out)
                                 cmd += " --save_model {}".format(model_filename)
                                 cmd += " --learning_rate {}".format(lr)
-                                cmd += " --number_blocks {}".format(32 * (256 / window))
+                                cmd += " --number_blocks {}".format(int(32 * (256 / window)))
                                 if i > 0:
                                     cmd += " --load_model {}".format(model_filename)
 
@@ -740,7 +740,7 @@ def main():
                                     cmd += " --window_length {}".format(window)
                                     cmd += " --input_file_swarm {}".format(original_swarm_file)
                                     cmd += " --save_file_samples {}".format(original_swarm_file_window)
-                                    cmd += " --number_blocks {}".format(32 * (256 / window))
+                                    cmd += " --number_blocks {}".format(int(32 * (256 / window)))
                                     run_cmd(cmd)
 
                                 if not check_files(["{}.npz".format(failed_swarm_file_window)]):
@@ -749,7 +749,7 @@ def main():
                                     cmd += " --window_length {}".format(window)
                                     cmd += " --input_file_swarm {}".format(failed_swarm_file)
                                     cmd += " --save_file_samples {}".format(failed_swarm_file_window)
-                                    cmd += " --number_blocks {}".format(32 * (256 / window))
+                                    cmd += " --number_blocks {}".format(int(32 * (256 / window)))
                                     run_cmd(cmd)
 
                                 cmd = "python3 main.py Predict"
@@ -759,7 +759,7 @@ def main():
                                 cmd += " --input_predict {}".format(failed_swarm_file_window)
                                 cmd += " --output_predict {}".format(corrected_swarm_file)
                                 cmd += " --load_model {}".format(model_filename)
-                                cmd += " --number_blocks {}".format(32 * (256 / window))
+                                cmd += " --number_blocks {}".format(int(32 * (256 / window)))
                                 run_cmd(cmd)
 
                                 time_end_experiment = datetime.datetime.now()
@@ -774,7 +774,7 @@ def main():
                                 cmd += " --epochs {}".format(NUM_EPOCHS)
                                 cmd += " --pif {}".format(pif)
                                 cmd += " --duration {}".format(duration)
-                                cmd += " --number_blocks {}".format(32 * (256 / window))
+                                cmd += " --number_blocks {}".format(int(32 * (256 / window)))
 
                                 cmd += " --file_original {}".format(original_swarm_file)
                                 cmd += " --file_corrected {}".format(corrected_swarm_file)
